Refactor: AccessToken 요청이 APIRequest를 사용하도록 변경
