+++ PRTracker/Source/Common/Networking/HTTPMethod.swift
@@ -1,0 +1,19 @@
+//
+//  HTTPMethod.swift
+//  PRTracker
+//
+//  Created by Bumgeun Song on 2022/06/27.
+//
+
+import Foundation
+
+enum HTTPMethod: String {
+    case get = "GET"
+    case post = "POST"
+    case patch = "PATCH"
+    case delete = "DELETE"
+}
+
+enum MediaType: String {
+    case json = "application/json"
+}
