+++ PRTracker/Source/Login/Model/GitHubLoginManager.swift
@@ -22,8 +22,7 @@ struct GitHubLoginManager {
     
     static let requiredScope = "repo,user"
     static let authorizeBaseURL = "https://github.com/login/oauth/authorize"
-    static let accessTokenURL = "https://github.com/login/oauth/access_token"
-    static let validateTokenURL = "https://github.com/applications/\(Secrets.clientId)/token"
+    
     static let keyChainAccount = "github"
     static let keyChainToken = "access-token"
     
@@ -37,10 +36,10 @@ struct GitHubLoginManager {
     }
     
     func requestAuthorization() {
-        guard var components = URLComponents(string: GitHubLoginManager.authorizeBaseURL) else { return }
+        guard var components = URLComponents(string: Self.authorizeBaseURL) else { return }
         components.queryItems = [
             URLQueryItem(name: "client_id", value: Secrets.clientId),
-            URLQueryItem(name: "scope", value: GitHubLoginManager.requiredScope)
+            URLQueryItem(name: "scope", value: Self.requiredScope)
         ]
         guard let url = components.url else { return }
         UIApplication.shared.open(url)
@@ -48,16 +47,14 @@ struct GitHubLoginManager {
     
     let authorization: Observable<AuthorizationStatus> = Observable(.none)
     
-    func getAccessToken(with code: String) {
-        guard let url = makeAccessTokenURL(with: code) else { return }
-        let request = makeAccessTokenRequest(with: url)
+    func requestAccessToken(with code: String) {
+        let tokenResource = TokenResource(clientID: Secrets.clientId, clientSecret: Secrets.clientSecret, code: code)
+        let tokenRequest = APIRequest(resource: tokenResource, httpMethod: .post)
         
-        networkService.request(request) { (result: Result<TokenResponse, NetworkError>) in
+        networkService.execute(tokenRequest) { result in
             switch result {
             case .success(let data):
-                keyChainService.save(data.accessToken,
-                                     service: GitHubLoginManager.keyChainToken,
-                                     account: GitHubLoginManager.keyChainAccount)
+                keyChainService.save(data.accessToken, service: Self.keyChainToken, account: Self.keyChainAccount)
                 authorization.value = .authorized
                 
             case .failure(let error):
@@ -73,22 +70,6 @@ struct GitHubLoginManager {
             if user == nil { return completion(false) }
             return completion(true)
         }
-    }
-    
-    private func makeAccessTokenURL(with code: String) -> URL? {
-        guard var components = URLComponents(string: GitHubLoginManager.accessTokenURL) else { return nil }
-        components.queryItems = [
-            URLQueryItem(name: "client_id", value: Secrets.clientId),
-            URLQueryItem(name: "client_secret", value: Secrets.clientSecret),
-            URLQueryItem(name: "code", value: code)
-        ]
-        return components.url
-    }
-    
-    private func makeAccessTokenRequest(with url: URL) -> URLRequest {
-        var request = URLRequest(url: url)
-        request.setValue("application/json", forHTTPHeaderField: "Accept")
-        return request
     }
 }
