+++ PRTracker/Source/PR_Listing/Model/IssuesResource.swift
@@ -7,6 +7,9 @@
 
 import Foundation
 
+
+
+
 struct IssuesResource: APIResource {
     typealias ModelType = [Issue]
