--- PRTracker/Source/PR_Listing/Model/ListingManager.swift
@@ -1,8 +1,0 @@
-//
-//  ListingManager.swift
-//  PRTracker
-//
-//  Created by Bumgeun Song on 2022/06/13.
-//
-
-import Foundation
