--- PRTracker/Source/PR_Listing/Model/PullListModel.swift
@@ -1,16 +1,0 @@
-//
-//  PullListModel.swift
-//  PRTracker
-//
-//  Created by 안상희 on 2022/06/17.
-//
-
-import Foundation
-
-struct PullListModel {
-    let title: String
-    let content: String
-    let projectName: String
-    let labelName: String?
-    let commentCount: Int?
-}
