+++ PRTracker/Source/PR_Listing/Model/UserManager.swift
@@ -27,7 +27,7 @@ struct UserManager {
         let userResource = UserResource()
         let userRequest = APIRequest(resource: userResource, token: accessToken)
         
-        userRequest.execute { result in
+        networkService.execute(userRequest) { result in
             switch result {
             case .success(let data):
                 completion(data)
